requirements.txt
torch
transformers 
requests
Pillow 
fastapi 
uvicorn[standard]
python-multipart